feat(upload folder): init manifest options PE-6636
